Добавлен функционал для машин и домов

+++ src/main/java/org/example/citydatabase/repository/HouseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HouseRepository extends CrudRepository<House, Integer> {
+public interface HouseRepository extends CrudRepository<House, Long> {
 }
